test: more fs stuff
